index优化

+++ src/main/java/com/ahuthj/enums/IsDeleteEnum.java
@@ -14,17 +14,17 @@ import lombok.Getter;
  * Created by Bryant on 2018.11.8
  */
 public enum IsDeleteEnum {
-    NOT_DELETE((byte)0, "未删除"),
-    NORMAL_DELETE((byte)1, "正常删除"),
-    UNSALABLE_DELETE((byte)2, "滞销删除");
+    NOT_DELETE(0, "未删除"),
+    NORMAL_DELETE(1, "正常删除"),
+    UNSALABLE_DELETE(2, "滞销删除");
 
     @Getter
-    private Byte code;
+    private Integer code;
 
     @Getter
     private String description;
 
-    IsDeleteEnum(Byte code, String description) {
+    IsDeleteEnum(Integer code, String description) {
         this.code = code;
         this.description = description;
     }
